Organize files and SPIFFS w BLE bug fixes
