+++ .vscode/settings.json
@@ -22,6 +22,7 @@
         "ratio": "c",
         "streambuf": "c",
         "tuple": "c",
-        "type_traits": "c"
+        "type_traits": "c",
+        "circular_buffer.h": "c"
     }
 }
